pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "vector-chat"
version = "0.1.0"
description = "A Python toolkit for text embedding with OpenAI and Qdrant, semantic search, and AI-powered chat"
authors = ["Josue Gutierrez <josue.ggh@gmail.com>"]
readme = "README.md"
packages = [{include = "vector_chat"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.0.0"
python-dotenv = "^1.0.0"
qdrant-client = "^1.6.0"
nltk = "^3.8.1"
requests = "^2.31.0"
numpy = "^1.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[tool.poetry.scripts]
embed = "vector_chat.cli.embed:main"
chat = "vector_chat.cli.chat:main"
vector-chat = "vector_chat.__main__:main"

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py" 